refactoring: rename ast::collections -> ast::subst_map
